refactor: introduce api common types (#3365)
* fix: configmap resource filters generated by helm does not account for namespace


* refactor: move api functions closer to the struct they belong to


* refactor: introduce api common types
